+++ packages/client/src/app/ui/protected-routes.pcss
@@ -8,7 +8,7 @@
   border-bottom: 1px solid #5f5959;
 }
 
-.content {
+.container {
   text-align: center;
   height: calc(100vh - 103px);
   padding-inline: 107px;
