+++ packages/client/src/assets/wide_logo.png
