--- packages/client/src/shared/config/variables.pcss
@@ -1,13 +1,0 @@
-:root {
-  /* Colors */
-  --auth-form-bg: #0D0F1C;
-  --auth-form-border: rgb(255 255 255 / 20%);
-  --auth-primary-button: #1677ff;
-  --auth-text-color: #fff;
-  
-  /* Font */
-  --auth-font-family: 'Nunito Sans', sans-serif;
-  
-  /* Sizes */
-  --input-height: 45px;
-}
