+++ packages/client/src/shared/ui/auth-form/auth-form.pcss
@@ -1,12 +1,3 @@
-.auth-form-container {
-  background-color: var(--auth-form-bg);
-  padding: 30px;
-  border-radius: 16px;
-  width: 100%;
-  max-width: 400px;
-  font-family: var(--auth-font-family);
-}
-
 .auth-form-title {
   font-weight: 700;
   font-size: 36px;
