--- packages/client/src/shared/ui/auth-layout/index.ts
@@ -1,1 +1,0 @@
-export { AuthLayout } from './auth-layout'
