+++ packages/client/src/shared/ui/index.ts
@@ -1,3 +1,3 @@
-export { AuthLayout } from './auth-layout'
+export { AuthLayout } from './layout/auth-layout'
 export { AuthForm } from './auth-form'
 export type { FormField } from './auth-form'
