+++ packages/client/src/shared/ui/layout/app-layout.module.pcss
@@ -1,0 +1,4 @@
+.layout {
+  height: 100vh;
+  overflow: auto;
+}
