+++ packages/client/src/shared/ui/layout/auth-layout.module.pcss
@@ -1,0 +1,22 @@
+.video-background {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  overflow: hidden;
+
+  video {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+  }
+}
+
+.container {
+  height: 100%;
+
+  .image {
+    max-width: 75%;
+  }
+}
